packs/elastic/requirements.txt
elasticsearch-curator==3.0.3
easydict
